took changes from NRMC26 that were directly pertinant to that branch

+++ src/controls/wheel_control/include/waypoint_controller/pose.h
@@ -3,7 +3,7 @@
 
 struct pose
 {
-  float x, y, theta;
+  double x, y, theta;
 };
 
 #endif
